fix: error build in job available
