+++ src/services/job.ts
@@ -35,21 +35,6 @@ export type TPaginatedJobs = {
   };
 };
 
-export function getRelativeTime(date: Date | string): string {
-  const d = typeof date === "string" ? new Date(date) : date;
-  const diffMs = Date.now() - d.getTime();
-  const diffSec = Math.floor(diffMs / 1000);
-  if (diffSec < 60) return "Baru saja";
-  const diffMin = Math.floor(diffSec / 60);
-  if (diffMin < 60) return `${diffMin} menit yang lalu`;
-  const diffHr = Math.floor(diffMin / 60);
-  if (diffHr < 24) return `${diffHr} jam yang lalu`;
-  const diffDay = Math.floor(diffHr / 24);
-  if (diffDay < 7) return `${diffDay} hari yang lalu`;
-  const diffWk = Math.floor(diffDay / 7);
-  return `${diffWk} minggu yang lalu`;
-}
-
 export const createJobService = async (data: TNewJob): Promise<TJob> => {
   try {
     const [job] = await db.insert(jobTable).values({
